+++ tests/lightning_tensor/test_quimb_mps.py
@@ -40,8 +40,8 @@ class TestQuimbMPS:
 
         wires = Wires(range(num_wires)) if num_wires else None
         dev = LightningTensor(wires=wires, backend=backend, method=method, c_dtype=c_dtype)
-        assert isinstance(dev.state, qtn.MatrixProductState)
-        assert isinstance(dev.state_to_array(), np.ndarray)
+        assert isinstance(dev._interface.state, qtn.MatrixProductState)
+        assert isinstance(dev._interface.state_to_array(), np.ndarray)
 
         config = dev.preprocess()
         assert config.device_options["backend"] == backend